# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mehnat.uz Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface UploadResult {
  filename: string;
  rowCount: number;
  message: string;
}

export type BotStatusStatus =
  (typeof BotStatusStatus)[keyof typeof BotStatusStatus];

export const BotStatusStatus = {
  idle: "idle",
  running: "running",
  stopping: "stopping",
  error: "error",
  completed: "completed",
} as const;

export interface BotStatus {
  status: BotStatusStatus;
  currentRunId: string | null;
  totalRows: number;
  processedRows: number;
  failedRows: number;
  startedAt: string | null;
  hasExcelFile: boolean;
}

export type BotRunStatus = (typeof BotRunStatus)[keyof typeof BotRunStatus];

export const BotRunStatus = {
  running: "running",
  completed: "completed",
  stopped: "stopped",
  error: "error",
} as const;

export interface BotRun {
  id: string;
  status: BotRunStatus;
  startedAt: string;
  completedAt: string | null;
  totalRows: number;
  processedRows: number;
  failedRows: number;
}

export type LogEntryLevel = (typeof LogEntryLevel)[keyof typeof LogEntryLevel];

export const LogEntryLevel = {
  info: "info",
  warn: "warn",
  error: "error",
  success: "success",
} as const;

export interface LogEntry {
  id: string;
  timestamp: string;
  level: LogEntryLevel;
  message: string;
  rowNumber: number | null;
}

export type ExcelPreviewRowsItem = { [key: string]: unknown };

export interface ExcelPreview {
  headers: string[];
  rows: ExcelPreviewRowsItem[];
  totalCount: number;
}

export interface BotSettings {
  password: string;
  employeeSearch: string;
  defaultHudud: string;
  defaultTuman: string;
  defaultMahalla: string;
  defaultYonalish: string;
  defaultOy: string;
  defaultShartnomaUri: string;
  defaultIshBoshlangan: string;
  defaultShartnamaSan: string;
  defaultBajarilgan: string;
  headless: boolean;
}

export type UploadExcelBody = {
  file: Blob;
};

export type GetBotLogsParams = {
  limit?: number;
};