# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Mehnat.uz Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/upload:
    post:
      operationId: uploadExcel
      tags: [bot]
      summary: Upload Excel file with employee data
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bot/start:
    post:
      operationId: startBot
      tags: [bot]
      summary: Start the bot
      responses:
        "200":
          description: Bot started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotRun"
        "400":
          description: Bot already running or no Excel file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /bot/logs:
    get:
      operationId: getBotLogs
      tags: [bot]
      summary: Get recent bot logs
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
          required: false
      responses:
        "200":
          description: Log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LogEntry"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UploadResult:
      type: object
      properties:
        filename:
          type: string
        rowCount:
          type: integer
        message:
          type: string
      required:
        - filename
        - rowCount
        - message
...
    BotRun:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [running, completed, stopped, error]
        startedAt:
          type: string
        completedAt:
          type: string
          nullable: true
        totalRows:
          type: integer
        processedRows:
          type: integer
        failedRows:
          type: integer
      required:
        - id
        - status
        - startedAt
        - completedAt
        - totalRows
        - processedRows
        - failedRows
    LogEntry:
      type: object
      properties:
        id:
          type: string
        timestamp:
          type: string
        level:
          type: string
          enum: [info, warn, error, success]
        message:
          type: string
        rowNumber:
          type: integer
          nullable: true
      required:
        - id
        - timestamp
        - level
        - message
        - rowNumber
...