# lib/api-zod/src/generated/types/botRun.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mehnat.uz Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { BotRunStatus } from "./botRunStatus";

export interface BotRun {
  id: string;
  status: BotRunStatus;
  startedAt: string;
  completedAt: string | null;
  totalRows: number;
  processedRows: number;
  failedRows: number;
}