# lib/api-zod/src/generated/types/logEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mehnat.uz Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { LogEntryLevel } from "./logEntryLevel";

export interface LogEntry {
  id: string;
  timestamp: string;
  level: LogEntryLevel;
  message: string;
  rowNumber: number | null;
}