# lib/api-zod/src/generated/types/uploadResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Mehnat.uz Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResult {
  filename: string;
  rowCount: number;
  message: string;
}